feat: 优化附加调试器时页面抖动问题，优化代码结构
